Texture2D

+++ include/Lumino/Animation/EasingFunctions.h
@@ -132,21 +132,21 @@ public:
 	template<typename TValue, typename TTime>
 	static TValue EaseInSine(TTime t, TValue b, TValue c, TTime d)
 	{
-		return -c * std::cos(t / d * (M_PI / 2.0)) + c + b;
+		return -c * std::cos(t / d * (((float)M_PI) / 2.0f)) + c + b;
 	};
 
 	/// sinusoidal easing out - decelerating to zero velocity
 	template<typename TValue, typename TTime>
 	static TValue EaseOutSine(TTime t, TValue b, TValue c, TTime d)
 	{
-		return c * std::sin(t / d * (M_PI / 2.0)) + b;
+		return c * std::sin(t / d * (((float)M_PI) / 2.0f)) + b;
 	};
 
 	/// sinusoidal easing in/out - accelerating until halfway, then decelerating
 	template<typename TValue, typename TTime>
 	static TValue EaseInOutSine(TTime t, TValue b, TValue c, TTime d)
 	{
-		return -c / 2.0 * (std::cos(M_PI*t / d) - 1) + b;
+		return -c / 2.0 * (std::cos(((float)M_PI) *t / d) - 1) + b;
 	};
 
 	/// exponential easing in - accelerating from zero velocity
